pyproject.toml
[project]
name = "binance-futures-bot"
version = "0.1.0"
description = "A simple trading bot for Binance Futures Testnet"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.0",
    "rich>=13.7.1",
    "questionary>=2.0.1"
]

[project.scripts]
bot = "cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
